re-split provenance from modal + use new provenance for datasets
